requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
confusable-homoglyphs==3.2.0
Django==4.0
djangorestframework==3.13.1
psycopg2==2.9.3
pytz==2022.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
